requirements.txt
Django==4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
psycopg2-binary==2.9.3
ipdb==0.13.9
pandas==1.5.1
odfpy==1.4.1
selenium==4.2.0
beautifulsoup4==4.11.1
requests==2.27.1
lxml==4.9.0
webdriver-manager==3.7.0
celery==5.2.7
redis==4.3.3
flower==1.0.0